# Bulk Lazy initialization of FunctionDefinitions 