--- a/internal/wasm/function_definition.go
+++ b/internal/wasm/function_definition.go
@@ -31,16 +31,19 @@ func (m *Module) ExportedFunctions() map[string]api.FunctionDefinition {
 
 // FunctionDefinition returns the FunctionDefinition for the given `index`.
 func (m *Module) FunctionDefinition(index Index) *FunctionDefinition {
-	// TODO: lazy initialization.
+	// TODO: lazy initialization per function.
+	if len(m.FunctionDefinitionSection) == 0 {
+		m.buildFunctionDefinitions()
+	}
 	return &m.FunctionDefinitionSection[index]
 }
 
-// BuildFunctionDefinitions generates function metadata that can be parsed from
+// buildFunctionDefinitions generates function metadata that can be parsed from
 // the module. This must be called after all validation.
 //
 // Note: This is exported for tests who don't use wazero.Runtime or
 // NewHostModule to compile the module.
-func (m *Module) BuildFunctionDefinitions() {
+func (m *Module) buildFunctionDefinitions() {
 	var moduleName string
 	var functionNames NameMap
 	var localNames, resultNames IndirectNameMap