`BuildFunctionDefinitions` is one of the most heavy allocation source during compilation,
which should be deferred when it's actually required (e.g. facing runtime errors).

This refactors the usage of FunctionDefinition, and defer the initialization of `[]FunctionDefinition`
to the time when `Module.FunctionDefinition` is called.

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
                                   │   old.txt   │              new.txt              │
                                   │   sec/op    │   sec/op     vs base              │
Compilation_sqlite3/compiler-10      134.4m ± 1%   134.3m ± 5%       ~ (p=0.620 n=7)
Compilation_sqlite3/interpreter-10   45.75m ± 0%   45.67m ± 1%       ~ (p=0.456 n=7)
geomean                              78.42m        78.30m       -0.16%

                                   │   old.txt    │              new.txt               │
                                   │     B/op     │     B/op      vs base              │
Compilation_sqlite3/compiler-10      54.32Mi ± 0%   54.04Mi ± 0%  -0.51% (p=0.001 n=7)
Compilation_sqlite3/interpreter-10   51.75Mi ± 0%   51.47Mi ± 0%  -0.54% (p=0.001 n=7)
geomean                              53.02Mi        52.74Mi       -0.52%

                                   │   old.txt   │              new.txt               │
                                   │  allocs/op  │  allocs/op   vs base               │
Compilation_sqlite3/compiler-10      26.06k ± 0%   23.00k ± 0%  -11.72% (p=0.001 n=7)
Compilation_sqlite3/interpreter-10   13.90k ± 0%   10.84k ± 0%  -22.00% (p=0.001 n=7)
geomean                              19.03k        15.79k       -17.02%

```