@mathetake 2023-05-02
decided to go ahead and do the minimal lazy initializtion. will push the commit in a bit