@mathetake 2023-05-02
maybe not worth a try considering the fact that there's no op/sec improvement