@evacchi 2023-05-08
I pushed a possible change (feel free to revert). If I am running the sqlite3 bench right, these are the results:

```
wazero on  lazyfunctiondefinition [$?⇡] via 🐹 v1.20.4
❯ go test -run='^$' -bench '^BenchmarkCompilation.*' ./internal/integration_test/bench -count=6 > new.txt

wazero on  lazyfunctiondefinition [$?⇡] via 🐹 v1.20.4 took 41s
❯ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
                                    │    old.txt    │               new.txt               │
                                    │    sec/op     │    sec/op      vs base              │
Compilation/with_extern_cache-10        81.90µ ± 2%     80.57µ ± 0%  -1.62% (p=0.002 n=6)
Compilation/without_extern_cache-10     991.4µ ± 0%     964.7µ ± 0%  -2.69% (p=0.002 n=6)
Compilation/interpreter-10              333.2µ ± 0%     327.0µ ± 1%  -1.86% (p=0.002 n=6)
Compilation_Sqlite3/interpreter-10    0.04533n ± 0%   0.04456n ± 0%  -1.70% (p=0.002 n=6)
Compilation_Sqlite3/compiler-10        0.1465n ± 1%    0.1427n ± 0%  -2.63% (p=0.002 n=6)
geomean                                 709.4n          694.5n       -2.10%
```