@mathetake 2023-05-09
```
Compilation/without_extern_cache-10     2.058m ± 1%     2.055m ± 1%       ~ (p=0.485 n=6)
Compilation/interpreter-10              638.5µ ± 1%     634.7µ ± 1%       ~ (p=0.065 n=6)
Compilation_Sqlite3/interpreter-10    0.04584n ± 1%   0.04559n ± 1%       ~ (p=0.180 n=6)
Compilation_Sqlite3/compiler-10        0.1443n ± 1%    0.1439n ± 1%       ~ (p=0.457 n=6)
```

this clearly indicates that you have a bug in your `Compilation_Sqlite3` - compiling sqlite3 must not be that fast and shouldn't be faster than compiling small tinygo binary in any way.