@mathetake 2023-05-09
on second thought, this might have a impact on listener heavy usecase, so I would like to request reviews from @achille-roussel and @pelletier