@achille-roussel 2023-05-09
It does seem to have a slight overhead:

```
$ go tool pprof -text -base /tmp/profile.0 /tmp/profile.1
Type: cpu
Time: May 9, 2023 at 3:55pm (PDT)
Duration: 10.06s, Total samples = 9.52s (94.60%)
Showing nodes accounting for -0.02s, 0.21% of 9.52s total
      flat  flat%   sum%        cum   cum%
    -0.24s  2.52%  2.52%     -0.24s  2.52%  runtime._ExternalCode
    -0.16s  1.68%  4.20%     -0.07s  0.74%  github.com/stealthrocket/wzprof.cpuProfiler.Before
     0.08s  0.84%  3.36%      0.15s  1.58%  github.com/tetratelabs/wazero/internal/engine/compiler.(*callEngine).builtinFunctionFunctionListenerAfter
     0.07s  0.74%  2.63%      0.07s  0.74%  sync.(*Mutex).Unlock (inline)
     0.06s  0.63%  2.00%      0.05s  0.53%  github.com/stealthrocket/wzprof.cpuProfiler.After
     0.05s  0.53%  1.47%      0.01s  0.11%  github.com/tetratelabs/wazero/internal/engine/compiler.(*callEngine).builtinFunctionFunctionListenerBefore
     0.05s  0.53%  0.95%      0.05s  0.53%  github.com/tetratelabs/wazero/internal/wasm.(*Module).FunctionDefinition
     0.04s  0.42%  0.53%      0.22s  2.31%  github.com/tetratelabs/wazero/internal/engine/compiler.(*callEngine).execWasmFunction
     0.04s  0.42%  0.11%      0.04s  0.42%  github.com/tetratelabs/wazero/internal/engine/compiler.(*stackIterator).Next
    -0.04s  0.42%  0.53%     -0.05s  0.53%  runtime.nanotime1
    -0.03s  0.32%  0.84%      0.02s  0.21%  github.com/tetratelabs/wazero/internal/engine/compiler.(*function).definition (inline)
     0.02s  0.21%  0.63%      0.02s  0.21%  github.com/stealthrocket/wzprof.stackTrace.bytes (inline)
     0.02s  0.21%  0.42%      0.02s  0.21%  github.com/tetratelabs/wazero/internal/engine/compiler.nativecall
     0.02s  0.21%  0.21%      0.02s  0.21%  runtime.mapaccess1_fast64
    -0.02s  0.21%  0.42%     -0.07s  0.74%  runtime.nanotime
    -0.01s  0.11%  0.53%     -0.01s  0.11%  aeshashbody
     0.01s  0.11%  0.42%      0.01s  0.11%  github.com/tetratelabs/wazero/internal/engine/compiler.(*stackIterator).Function
     0.01s  0.11%  0.32%      0.01s  0.11%  github.com/tetratelabs/wazero/internal/engine/compiler.(*stackIterator).ProgramCounter
     0.01s  0.11%  0.21%      0.01s  0.11%  github.com/tetratelabs/wazero/internal/engine/compiler.(*stackIterator).clear (inline)
     0.01s  0.11%  0.11%      0.01s  0.11%  github.com/tetratelabs/wazero/internal/engine/compiler.(*stackIterator).reset (inline)
     0.01s  0.11%     0%      0.01s  0.11%  runtime.execute
    -0.01s  0.11%  0.11%     -0.01s  0.11%  runtime.libcCall
    -0.01s  0.11%  0.21%     -0.01s  0.11%  runtime.memhash
         0     0%  0.21%      0.06s  0.63%  github.com/stealthrocket/wzprof.makeStackTrace
         0     0%  0.21%      0.02s  0.21%  github.com/stealthrocket/wzprof.stackCounterMap.lookup (inline)
         0     0%  0.21%      0.02s  0.21%  github.com/stealthrocket/wzprof.stackCounterMap.observe
         0     0%  0.21%      0.22s  2.31%  github.com/tetratelabs/wazero.(*runtime).InstantiateModule
         0     0%  0.21%      0.22s  2.31%  github.com/tetratelabs/wazero/internal/engine/compiler.(*callEngine).Call
         0     0%  0.21%      0.22s  2.31%  github.com/tetratelabs/wazero/internal/engine/compiler.(*callEngine).call
         0     0%  0.21%     -0.02s  0.21%  hash/maphash.Bytes
         0     0%  0.21%     -0.02s  0.21%  hash/maphash.rthash (inline)
         0     0%  0.21%      0.22s  2.31%  main.(*program).run.func4
         0     0%  0.21%     -0.24s  2.52%  runtime._System
         0     0%  0.21%      0.01s  0.11%  runtime.gopreempt_m
         0     0%  0.21%      0.01s  0.11%  runtime.goschedImpl
         0     0%  0.21%      0.01s  0.11%  runtime.newstack
         0     0%  0.21%      0.01s  0.11%  runtime.schedule
```

This is the most impacted, almost twice more time is spent in this function:
```
     0.08s  0.84%  3.36%      0.15s  1.58%  github.com/tetratelabs/wazero/internal/engine/compiler.(*callEngine).builtinFunctionFunctionListenerAfter
```