@achille-roussel 2023-05-09
☝️ if this is a regression it's minor, maybe not worth worrying about it, we can always optimize later?