@evacchi 2023-05-15
thanks I didn't realize you already merged this. For reference, I have just re-run the sqlite3 benchmark, which should be correct now

```
❯ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
                                            │   old.txt   │              new.txt              │
                                            │   sec/op    │   sec/op     vs base              │
Compilation/with_extern_cache-10              169.6µ ± 2%   160.7µ ± 1%  -5.24% (p=0.002 n=6)
Compilation/without_extern_cache-10           2.053m ± 1%   2.056m ± 2%       ~ (p=0.589 n=6)
Compilation/interpreter-10                    635.9µ ± 1%   632.8µ ± 1%       ~ (p=0.132 n=6)
Compilation_Sqlite3/with_extern_cache-10      13.16m ± 0%   13.00m ± 0%  -1.26% (p=0.002 n=6)
Compilation_Sqlite3/without_extern_cache-10   145.2m ± 1%   145.1m ± 1%       ~ (p=0.310 n=6)
Compilation_Sqlite3/interpreter-10            46.94m ± 3%   46.76m ± 2%       ~ (p=0.132 n=6)
geomean                                       5.204m        5.140m       -1.24%

                                            │   old.txt    │               new.txt               │
                                            │     B/op     │     B/op      vs base               │
Compilation/with_extern_cache-10              64.94Ki ± 0%   49.02Ki ± 0%  -24.52% (p=0.002 n=6)
Compilation/without_extern_cache-10           1.103Mi ± 0%   1.088Mi ± 0%   -1.40% (p=0.002 n=6)
Compilation/interpreter-10                    774.8Ki ± 0%   759.0Ki ± 0%   -2.04% (p=0.002 n=6)
Compilation_Sqlite3/with_extern_cache-10      3.230Mi ± 0%   2.893Mi ± 0%  -10.43% (p=0.002 n=6)
Compilation_Sqlite3/without_extern_cache-10   76.12Mi ± 0%   75.80Mi ± 0%   -0.42% (p=0.002 n=6)
Compilation_Sqlite3/interpreter-10            57.25Mi ± 0%   56.93Mi ± 0%   -0.57% (p=0.002 n=6)
geomean                                       3.011Mi        2.800Mi        -7.01%

                                            │   old.txt    │              new.txt               │
                                            │  allocs/op   │  allocs/op   vs base               │
Compilation/with_extern_cache-10                582.0 ± 0%    420.0 ± 0%  -27.84% (p=0.002 n=6)
Compilation/without_extern_cache-10            1145.0 ± 0%    983.0 ± 0%  -14.15% (p=0.002 n=6)
Compilation/interpreter-10                      788.0 ± 0%    626.0 ± 0%  -20.56% (p=0.002 n=6)
Compilation_Sqlite3/with_extern_cache-10      12.787k ± 0%   9.259k ± 0%  -27.59% (p=0.002 n=6)
Compilation_Sqlite3/without_extern_cache-10    29.93k ± 0%   26.40k ± 0%  -11.79% (p=0.002 n=6)
Compilation_Sqlite3/interpreter-10             16.94k ± 0%   13.41k ± 0%  -20.81% (p=0.002 n=6)
geomean                                        3.879k        3.076k       -20.69%```